Update st7789.c
This fix allows displays with resolutions higher than height*width > 65535

+++ components/st7789/st7789.c
@@ -1034,13 +1034,13 @@ void lcdDrawFinish(TFT_t *dev)
 	spi_master_write_addr(dev, dev->_offsety, dev->_offsety+dev->_height-1);
 	spi_master_write_command(dev, 0x2C); // Memory Write
 
-	uint16_t size = dev->_width*dev->_height;
+	uint32_t size = dev->_width*dev->_height;
 	uint16_t *image = dev->_frame_buffer;
 	while (size > 0) {
 		// 1024 bytes per time.
 		uint16_t bs = (size > 1024) ? 1024 : size;
 		spi_master_write_colors(dev, image, bs);
-		size -= bs;
+		size -= (uint32_t)bs;
 		image += bs;
 	}
 	return;
